requirements.txt
asgiref==3.8.1
crispy-tailwind==1.0.3
Django==5.1.1
django-admin-interface==0.28.9
django-colorfield==0.11.0
django-crispy-forms==2.3
pillow==10.4.0
python-decouple==3.8
python-slugify==8.0.4
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
